+++ lib/Xtc/Xtc/XtcTypes.h
@@ -86,15 +86,15 @@ struct XtgPageHeader {
 };
 #pragma pack(pop)
 
-// Page information (internal use, optimized for memory)
+// Page information (internal use)
 struct PageInfo {
-  uint32_t offset;   // File offset to page data (max 4GB file size)
+  uint64_t offset;   // File offset to page data (64-bit to support files >4GB)
   uint32_t size;     // Data size (bytes)
   uint16_t width;    // Page width
   uint16_t height;   // Page height
   uint8_t bitDepth;  // 1 = XTG (1-bit), 2 = XTH (2-bit grayscale)
   uint8_t padding;   // Alignment padding
-};  // 16 bytes total
+};  // 20 bytes total
 
 struct ChapterInfo {
   std::string name;
